requirements.txt
python-dotenv
tqdm
requests
joblib
numpy
pandas
regex
emoji
scipy
scikit-learn
torch
nltk
matplotlib
seaborn
umap-learn
mlflow
openpyxl
jupyter
ipykernel
apache-airflow
pytest
pytest-cov

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

mail-parser
beautifulsoup4
readability-lxml
ftfy
python-dateutil
email-reply-parser

langchain-google-community
langchain-community
sentence-transformers>=2.7.0
transformers > 4.51.0
huggingface-hub
accelerate
sentencpiece
safetensors
pyhercules
google-genai








